+++ src/components/tooltip/tooltip.styles.ts
@@ -18,6 +18,7 @@ export default css`
   }
 
   .tooltip::part(popup) {
+    pointer-events: none;
     z-index: var(--sl-z-index-tooltip);
   }